# Improve performance of parsing repeated fixed sized types