Based on https://github.com/protocolbuffers/protobuf/pull/7351. Will rebase on master once it is merged.

Actual changes: https://github.com/JamesNK/protobuf/commit/d0d576e316a0944bc6eb2f38b371ccc2ddce15c8

@jtattermusch 

Before:
```
|                                                          Method | messageCount |        Mean |      Error |     StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|---------------------------------------------------------------- |------------- |------------:|-----------:|-----------:|------------:|------------:|------------:|--------------------:|
|                         RepeatedFieldMessage_ParseFromByteArray |            ? |    13.44 us |  0.1718 us |  0.1523 us |     21.4996 |           - |           - |            16.54 KB |
|                  RepeatedFieldMessage_ParseFromReadOnlySequence |            ? |    13.11 us |  0.0688 us |  0.0610 us |     21.2708 |           - |           - |            16.36 KB |
|        RepeatedFieldMessage_ParseDelimitedMessagesFromByteArray |           10 |   133.31 us |  1.8114 us |  1.6944 us |    212.6465 |           - |           - |           163.77 KB |
| RepeatedFieldMessage_ParseDelimitedMessagesFromReadOnlySequence |           10 |   126.78 us |  0.8992 us |  0.7971 us |    212.6465 |           - |           - |           163.59 KB |
|        RepeatedFieldMessage_ParseDelimitedMessagesFromByteArray |          100 | 1,332.55 us | 28.6365 us | 25.3855 us |   2126.9531 |           - |           - |          1636.12 KB |
| RepeatedFieldMessage_ParseDelimitedMessagesFromReadOnlySequence |          100 | 1,273.07 us | 13.5838 us | 10.6054 us |   2126.9531 |           - |           - |          1635.94 KB |
```
After:
```
|                                                          Method | messageCount |       Mean |      Error |     StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|---------------------------------------------------------------- |------------- |-----------:|-----------:|-----------:|------------:|------------:|------------:|--------------------:|
|                         RepeatedFieldMessage_ParseFromByteArray |            ? |   6.127 us |  0.0373 us |  0.0330 us |     10.6354 |           - |           - |             8.25 KB |
|                  RepeatedFieldMessage_ParseFromReadOnlySequence |            ? |   6.095 us |  0.0695 us |  0.0581 us |     10.4141 |           - |           - |             8.07 KB |
|        RepeatedFieldMessage_ParseDelimitedMessagesFromByteArray |           10 |  60.599 us |  0.5330 us |  0.4725 us |    105.2246 |           - |           - |            80.88 KB |
| RepeatedFieldMessage_ParseDelimitedMessagesFromReadOnlySequence |           10 |  59.789 us |  0.2847 us |  0.2377 us |    104.1260 |           - |           - |             80.7 KB |
|        RepeatedFieldMessage_ParseDelimitedMessagesFromByteArray |          100 | 622.160 us | 12.3938 us | 20.7072 us |   1041.0156 |           - |           - |           807.21 KB |
| RepeatedFieldMessage_ParseDelimitedMessagesFromReadOnlySequence |          100 | 597.545 us |  5.5646 us |  4.9328 us |   1041.0156 |           - |           - |           807.03 KB |
```