@JamesNK 2020-06-02
Mergable now that https://github.com/protocolbuffers/protobuf/pull/7490 is resolved? @jtattermusch